requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2-binary==2.9.6
Pillow==9.5.0
django-cors-headers==4.1.0
djangorestframework-simplejwt==5.2.2